refactor storybook Unit component tests
